fix(server): 将 points 计数改为 long

+++ server/src/main/kotlin/infra/web/datasource/providers/Alphapolis.kt
@@ -100,7 +100,7 @@ class Alphapolis(
             .text()
             .substringBefore("pt")
             .filter { it.isDigit() }
-            .toInt()
+            .toLong()
 
         val totalCharacters = row("文字数")
             .text()
